+++ src/main/resources/day01/input.txt
