+++ src/main/resources/day02/input.txt
