+++ src/main/resources/day03/input.txt
