+++ src/main/resources/day04/input.txt
