+++ src/main/resources/day05/input.txt
