+++ src/main/resources/day06/input.txt
