+++ src/main/resources/day07/input.txt
