+++ src/main/resources/day08/input.txt
